fix disappearing content after refresh #871
Fix problem with `echo(..., {newline: false, flush: false})` where two lines ended
in the same container and the index came from the last line. When refreshing
the first line was ignored becuase the index in DOM was missing.
The fix was to reset the state after newline so both lines were in different
containers